chore(heartbeat): extract publicId and config-error helpers
- Add getHeartbeatPublicId to centralize dual-token publicId parsing.
- Add getConfigEvaluationHeartbeatErrors to share the fallback config
  evaluation error used by adapter heartbeat catch blocks.
- Refactor all Cloudflare adapters and the Vercel runner to use the new
  helpers, removing duplicated parseApiToken/sanitizeConfigErrors logic.
- Keep sanitizeConfigErrors calls in the adapters so existing safety tests
  that spy on the barrel export continue to work.
